+++ packages/site/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [19.0.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.4...@pluralsight/ps-design-system-site@19.0.5) (2020-07-15)
+
+
+### Bug Fixes
+
+* **position:** add yOffset to clipAdjust ([#1051](https://github.com/pluralsight/design-system/issues/1051)) ([5fed4c2](https://github.com/pluralsight/design-system/commit/5fed4c24b6eb710a03648497738d5e9aef4337f2))
+
+
+
+
+
+## [19.0.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.3...@pluralsight/ps-design-system-site@19.0.4) (2020-07-03)
+
+
+### Bug Fixes
+
+* **table:** doubled up click handler nulling sort ([#1049](https://github.com/pluralsight/design-system/issues/1049)) ([2b02663](https://github.com/pluralsight/design-system/commit/2b0266309a7a4b673940eed21a1c4b74e960ca11))
+
+
+
+
+
+## [19.0.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.2...@pluralsight/ps-design-system-site@19.0.3) (2020-07-02)
+
+
+### Bug Fixes
+
+* **site:** typo ([c0324c2](https://github.com/pluralsight/design-system/commit/c0324c29a91cae43b985c0196afa0db0296f4106))
+
+
+
+
+
 ## [19.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@19.0.1...@pluralsight/ps-design-system-site@19.0.2) (2020-06-25)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-site