+++ packages/typeahead/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@4.0.17...@pluralsight/ps-design-system-typeahead@4.0.18) (2020-07-15)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
+## [4.0.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@4.0.16...@pluralsight/ps-design-system-typeahead@4.0.17) (2020-07-02)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-typeahead
+
+
+
+
+
 ## [4.0.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-typeahead@4.0.15...@pluralsight/ps-design-system-typeahead@4.0.16) (2020-06-25)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-typeahead